backend/ai_service/requirements.txt
textblob==0.17.1
mysql-connector-python==8.2.0
schedule==1.2.1
flask==3.0.0
flask-cors==4.0.0
